requirements.txt
pyspark==3.2.1
python-dotenv==0.20.0
requests==2.27.1
requests_oauthlib==1.3.1
